requirements.txt
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2
python-decouple
redis
requests
pillow

